Bugs #337 (new menu item), #464 (link shortening)
Bug #337: The "Events" drop-down menu will now have a new item that
links to the page that lets the user create a new event. This menu item
appears only when the user is logged in.

Bug #464: When sharing an event through twitter or facebook, the event
url is now shortened through bitly.
- You'll need to create two environment variables: BITLY_USERNAME,
BITLY_APIKEY